Improve files structure inside the package to follow duck pattern
