--- packages/seo-integration/src/analysis-result-containers/index.js
@@ -1,2 +1,0 @@
-export { default as ReadabilityResultsContainer } from "./readability";
-export { default as SeoResultsContainer } from "./seo";
