+++ packages/seo-integration/src/analysis-results/containers/index.js
@@ -1,0 +1,2 @@
+export { default as ReadabilityResultsContainer } from "./readability-container";
+export { default as SeoResultsContainer } from "./seo-container";
