+++ packages/seo-integration/src/analysis-results/containers/seo-container.js
@@ -3,8 +3,8 @@ import { useMemo } from "@wordpress/element";
 import { SEO_STORE_NAME } from "@yoast/seo-store";
 import { FOCUS_KEYPHRASE_ID } from "@yoast/seo-store";
 import { PropTypes } from "prop-types";
-import { useMarker } from "./hooks";
-import { transformAnalysisResults } from "./transformers";
+import { useMarker } from "../../analysis-result-containers/hooks";
+import { transformAnalysisResults } from "../../analysis-result-containers/transformers";
 
 /**
  * Handles known data for a SEO results component.
