+++ packages/seo-integration/src/analysis-results/helpers/transformAnalysisResults.js
